# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hermes-own API",
    "description": "Hermes-own API — chat sessions, scheduled jobs, platform channels, model management, skills, memory, logs, file browser, group chat, and terminal.",
    "version": "0.6.25"
  },
  "servers": [
    {
      "url": "http://localhost:8648",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/hermes/sessions/hermes": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "Get hermes",
        "description": "GET /api/hermes/sessions/hermes",
        "operationId": "listHermesSessions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/hermes/sessions/hermes/{id}": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "Get :id",
        "description": "GET /api/hermes/sessions/hermes/:id",
        "operationId": "getHermesSession",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Token"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing authentication token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
...
    }
  }
}